Issue #1137 [com-archive-browser] Refactoring + parameter commands
https://gitlab.com/esa/NMF/nmf-issues/-/issues/1137
