+++ sdk/tools/com-archive-tool/src/main/java/esa/mo/nmf/com_archive_browser/adapters/ArchiveToRolesAdapter.java
@@ -1,17 +1,24 @@
-//------------------------------------------------------------------------------
-//
-// System : ccsds-common
-//
-// Sub-System : esa.mo.nmf.log_browser.adapters
-//
-// File Name : ArchiveToRolesAdapter.java
-//
-// Author : marcel.mikolajko
-//
-// Creation Date : 10.08.2021
-//
-//------------------------------------------------------------------------------
-package esa.mo.nmf.log_browser.adapters;
+/* ----------------------------------------------------------------------------
+ * Copyright (C) 2021      European Space Agency
+ *                         European Space Operations Centre
+ *                         Darmstadt
+ *                         Germany
+ * ----------------------------------------------------------------------------
+ * System                : ESA NanoSat MO Framework
+ * ----------------------------------------------------------------------------
+ * Licensed under European Space Agency Public License (ESA-PL) Weak Copyleft – v2.4
+ * You may not use this file except in compliance with the License.
+ *
+ * Except as expressly set forth in this License, the Software is provided to
+ * You on an "as is" basis and without warranties of any kind, including without
+ * limitation merchantability, fitness for a particular purpose, absence of
+ * defects or errors, accuracy or non-infringement of intellectual property rights.
+ *
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ----------------------------------------------------------------------------
+ */
+package esa.mo.nmf.com_archive_browser.adapters;
 
 import org.ccsds.moims.mo.com.archive.consumer.ArchiveAdapter;
 import org.ccsds.moims.mo.com.archive.structures.ArchiveDetailsList;
@@ -72,4 +79,3 @@ public class ArchiveToRolesAdapter extends ArchiveAdapter implements QueryStatus
         return rolesNames;
     }
 }
-//------------------------------------------------------------------------------
